addUseHibernateToSpring

+++ src/main/java/accident/repository/AccidentMem.java
@@ -5,12 +5,13 @@ import org.springframework.stereotype.Repository;
 
 import java.util.HashMap;
 
+
 @Repository
 public class AccidentMem {
 
-    private final HashMap<Integer, Accident> accidents = new HashMap<>();
+    private final HashMap<Long, Accident> accidents = new HashMap<>();
 
-    public HashMap<Integer, Accident> getAccidents() {
+    public HashMap<Long, Accident> getAccidents() {
         return accidents;
     }
 }
